requirements.txt
opencv-python
numpy
torch
torchvision
transformers
pillow
psutil
scipy
jiwer
sentencepiece
protobuf
accelerate
